max/pipelines/requirements.txt
# This was deleted as part of pipeline refactor
# https://github.com/modularml/modular/commit/d7e54f4a11b8f9a29e883d204dcacea90513db82#diff-b47636dfb4ba8723a713bfc288fdb324fa6e048067c7bbad0195795583edb21e
# But this is needed by benchmarking to setup the environment
# Adding this back for now to unblock.  
click >= 8.1.7
fastapi >= 0.115.3
gguf >= 0.10.0
requests >= 2.32.3
tokenizers >= 0.19.1
# PyTorch 2.5.1 with cuda support
torch == 2.5.1
torchvision == 0.20.1
torchaudio == 2.5.1
transformers >= 4.44.2
safetensors >= 0.4.5
psutil

# serve
hf-transfer>=0.1.9
httpx>=0.28.1,<0.29
huggingface-hub>=0.24.0
opentelemetry-api>=1.29.0
opentelemetry-exporter-otlp-proto-http>=1.27.0
opentelemetry-exporter-prometheus>=0.50b0
opentelemetry-sdk>=1.29.0,<2.0
pillow>=10.3.0
prometheus-client>=0.21.0
pydantic-settings>=2.7.1
pydantic
pyinstrument>=5.0.1
python-json-logger>=2.0.7
sse-starlette>=2.1.2
starlette==0.41.2 # transitively included by sse-starlette, >=0.40.0 is required for resolving CVE-2024-47874, 0.41.3 introduces a breaking change against mypy (https://github.com/encode/starlette/discussions/2757)
tokenizers>=0.19.0
uvicorn>=0.34.0
uvloop>=0.21.0
